Support injecting extra passport clients via override.php
Summary:
So far we have explicitly listed all OAUTH clients, with dedicated
configuration variables. This does not work when integrating external
components, so as a more flexible meachnism we pick up on
auth.extra_passport_clients, and allow injecting that value via the
config/override.php mechanism (which is the same as used in
AppServiceProvider.php).

Differential Revision: https://git.kolab.org/D5616