Make the greylist variables a bit more forgiving.
Summary:
Allow for quicker retries, and don't forget previous entries so quickly.

The minimum age protects about a spammer quickly trying again in
quick succession. A one minute delay should be enough to make it
considerably more expensive to get through.

We forget old information to deal with domain reuse, so holding on to
the data for considerably longer than 7 days seems fine as long as we
eventually expire.

Differential Revision: https://git.kolab.org/D5757